Implement DDC URI/notation mapping with exception
Notations of table classes start with a T that is not part of the URI so
the mapping between DDC URIs and DDC notations requires an exception.

This commit implements the mapping and adds addition of URIs to the
conversion so all DDC classes have both notation and URI. We might
remove the URI to save space in the database.